+++ src/framework/piral-base/src/system.ts
@@ -73,6 +73,10 @@ export function registerModule(name: string, resolve: ModuleResolver) {
         if (typeof content === 'function') {
           _exports('__esModule', true);
           _exports('default', content);
+        } else if (typeof content === 'object') {
+          if (content && !Array.isArray(content) && !('default' in content)) {
+            _exports('default', content);
+          }
         }
       }
     },